backend/requirements.txt
# BnB SmartChoice DSS Backend Requirements
# Python 3.9+

# Database
psycopg2-binary==2.9.9

# Environment Variables
python-dotenv==1.0.0

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.25.0
pydantic==2.5.3

# Data Processing & Scientific Computing
numpy==1.26.3
pandas==2.1.4
scipy==1.11.4

# CORS & HTTP
python-multipart==0.0.6

# Optional: Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional: Logging & Monitoring
python-json-logger==2.0.7

